// @types/generated/contentful.d.ts
export interface IEcosystemSectionFields {
  /** title field f.ex h2 tag type */
  ecosystemTitle: string;

  /** text advantege content 1 */
  textAdvantegeContent1: string;

  /** text advantege content 2 */
  textAdvantegeContent2: string;

  /** text advantege content 3 */
  textAdvantegeContent3: string;

  /** text advantege content 4 */
  textAdvantegeContent4: string;

  /** text advantege content 5 */
  textAdvantegeContent5: string;

  /** text advantege content 6 */
  textAdvantegeContent6: string;

  /** text advantege content 7 */
  textAdvantegeContent7: string;

  /** text advantege content 8 */
  textAdvantegeContent8: string;

  /** image belong to advntages */
  imageBelongToAdvntages: Asset;

  /** extra asset img 1 */
  extraAssetImg1: Asset;

  /** extra asset img 2 */
  extraAssetImg2: Asset;

  /** NAME OF THIS COLLECTION */
  ecosystemSectionFieldsCollection?: string | undefined;
}

/** Ecosystem section: with listed advantages/ benefits */

export interface IEcosystemSection extends Entry<IEcosystemSectionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "ecosystemSection";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}